requirements.txt
langchain
openai
trafilatura
streamlit
python-dotenv